+++ tests/unit/test_adaptivity_parallel.py
@@ -31,7 +31,7 @@ class TestGlobalAdaptivity(TestCase):
         configurator = MagicMock()
         configurator.get_adaptivity_similarity_measure = MagicMock(return_value="L1")
         adaptivity_controller = GlobalAdaptivityCalculator(
-            configurator, MagicMock(), 5, global_ids, rank=self._rank, comm=self._comm
+            configurator, 5, global_ids, rank=self._rank, comm=self._comm
         )
 
         # Force the activation of sim #0 and #4
@@ -105,7 +105,7 @@ class TestGlobalAdaptivity(TestCase):
         configurator.get_adaptivity_coarsening_const = MagicMock(return_value=0.2)
         configurator.get_adaptivity_similarity_measure = MagicMock(return_value="L2rel")
         adaptivity_controller = GlobalAdaptivityCalculator(
-            configurator, MagicMock(), 5, global_ids, rank=self._rank, comm=self._comm
+            configurator, 5, global_ids, rank=self._rank, comm=self._comm
         )
 
         adaptivity_controller._adaptivity_data_names = {
@@ -173,7 +173,7 @@ class TestGlobalAdaptivity(TestCase):
         configurator = MagicMock()
         configurator.get_adaptivity_similarity_measure = MagicMock(return_value="L1")
         adaptivity_controller = GlobalAdaptivityCalculator(
-            configurator, MagicMock(), 5, global_ids, rank=self._rank, comm=self._comm
+            configurator, 5, global_ids, rank=self._rank, comm=self._comm
         )
 
         adaptivity_controller.communicate_micro_output(