+++ tests/unit/test_micro_manager.py
@@ -119,7 +119,7 @@ class TestFunctioncalls(TestCase):
         """
         Test if the functions in the Config class work.
         """
-        config = micro_manager.Config(MagicMock(), "micro-manager-config.json")
+        config = micro_manager.Config("micro-manager-config.json")
         config.read_json_micro_manager()
         self.assertEqual(config._config_file_name.split("/")[-1], "dummy-config.xml")
         self.assertEqual(config._micro_file_name, "test_micro_manager")