+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## latest
 
+- Improve logging by wrapping Python logger in a class https://github.com/precice/micro-manager/pull/133
 - Add information about adaptivity tuning parameters https://github.com/precice/micro-manager/pull/131
 - Put computation of counting active steps inside the adaptivity variant `if` condition https://github.com/precice/micro-manager/pull/130